Unified: library dtypes

+++ src/sampletones/timers/lfsr.py
@@ -77,7 +77,7 @@ class LFSRTimer(Timer):
         lfsrs = self.lfsr_tables[self.short].lfsrs
         cumsum_table = self.lfsr_tables[self.short].cumsums.astype(np.float32)
 
-        indices = np.arange(self.frame_length + 1)
+        indices = np.arange(self.frame_length + 1, dtype=np.float32)
         delta = self._clocks_per_sample
         clock = indices * delta + self.clock
         changes = np.abs(np.diff(np.floor(clock)).astype(int))
@@ -94,7 +94,7 @@ class LFSRTimer(Timer):
 
         mask = pairs[:, 1] > pairs[:, 0]
         nonzero_pairs = pairs[mask]
-        means = np.array([np.mean(cumsum_table[start:end]) for start, end in nonzero_pairs])
+        means = np.array([np.mean(cumsum_table[start:end]) for start, end in nonzero_pairs], dtype=np.float32)
 
         differences[mask] = np.diff(np.concatenate([[0], means]))
         frame = 2.0 * np.cumsum(differences) - 1.0
@@ -177,8 +177,8 @@ class LFSRTimer(Timer):
         clocks_per_sample = self.calculate_clocks_per_sample(MAX_PERIOD)
         repeats = int(np.ceil(clocks_per_sample * self.frame_length / self.lfsr_period)) * 2 + 1
 
-        lfsrs = np.ones(MAX_LFSR, dtype=np.int64)
-        lfsr_to_index = -np.ones(MAX_LFSR + 1, dtype=np.int64)
+        lfsrs = np.ones(MAX_LFSR, dtype=np.int16)
+        lfsr_to_index = -np.ones(MAX_LFSR + 1, dtype=np.int16)
 
         lfsr = 1
         for i in range(MAX_LFSR):
@@ -192,7 +192,7 @@ class LFSRTimer(Timer):
             lfsr = forward_lfsr
 
         lfsrs = np.tile(lfsrs, repeats)
-        cumsums = np.concatenate([[0], np.cumsum(lfsrs, dtype=np.int64)]) & 1
+        cumsums = np.concatenate([[0], np.cumsum(lfsrs, dtype=np.int16)]) & 1
 
         return LFSRTTables(
             lfsr_to_index,
